Merge development into master: Release v1.0.0

+++ src/main/java/apiCRUD/apiSkinVideogame/repository/SkinRepository.java
@@ -1,9 +1,11 @@
 package apiCRUD.apiSkinVideogame.repository;
 
 import apiCRUD.apiSkinVideogame.model.Skin;
+import apiCRUD.apiSkinVideogame.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SkinRepository extends JpaRepository <Skin, Integer> {
+public interface SkinRepository extends JpaRepository<Skin, Long> {
+
 }
